# Conflicts:
#	backend/baton/src/docs/asciidoc/RunnerPostCreateReadApi.adoc
#	backend/baton/src/main/java/touch/baton/domain/runnerpost/controller/RunnerPostController.java
#	backend/baton/src/main/java/touch/baton/domain/runnerpost/service/RunnerPostService.java
#	backend/baton/src/main/java/touch/baton/domain/runnerpost/service/dto/RunnerPostApplicantCreateRequest.java
#	backend/baton/src/main/java/touch/baton/domain/supporter/repository/SupporterRunnerPostRepository.java
#	backend/baton/src/test/java/touch/baton/assure/common/AssuredSupport.java
#	backend/baton/src/test/java/touch/baton/assure/runnerpost/RunnerPostAssuredCreateTest.java
#	backend/baton/src/test/java/touch/baton/assure/runnerpost/RunnerPostAssuredSupport.java
#	backend/baton/src/test/java/touch/baton/document/runnerpost/create/RunnerPostCreateApiTest.java
#	backend/baton/src/test/java/touch/baton/domain/runnerpost/service/RunnerPostServiceCreateTest.java
#	backend/baton/src/test/java/touch/baton/domain/supporter/repository/SupporterRunnerPostRepositoryTest.java

+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/controller/response/RunnerPostResponse.java
@@ -126,7 +126,7 @@ public record RunnerPostResponse() {
                                         String reviewStatus
     ) {
 
-        public static ReferencedBySupporter of(final RunnerPost runnerPost, final int applicantCount) {
+        public static ReferencedBySupporter of(final RunnerPost runnerPost, final long applicantCount) {
             return new ReferencedBySupporter(
                     runnerPost.getId(),
                     runnerPost.getTitle().getValue(),
